# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stillwater API
  version: 0.25.0
  description: Artist metadata and image management for Emby, Jellyfin, and Kodi.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:1973/api/v1
    description: Local development
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      description: API token (sw_ prefix) via Authorization header
    apiKeyAuth:
      type: apiKey
      in: query
      name: apikey
      description: "Deprecated: prefer bearerAuth (Authorization header). Query-string tokens risk leakage via logs and referrers."
      x-deprecated: true
  schemas:
...
    Artist:
      type: object
      description: A music artist or group with full metadata.
      properties:
        id:
          type: string
        name:
          type: string
        sort_name:
          type: string
        type:
          type: string
          description: Artist type, e.g. "person" or "group".
        gender:
          type: string
        disambiguation:
          type: string
        musicbrainz_id:
          type: string
        audiodb_id:
          type: string
        discogs_id:
          type: string
        wikidata_id:
          type: string
        deezer_id:
          type: string
        spotify_id:
          type: string
        genres:
          type:
            - array
            - "null"
          items:
            type: string
        styles:
          type:
            - array
            - "null"
          items:
            type: string
        moods:
          type:
            - array
            - "null"
          items:
            type: string
        years_active:
          type: string
        born:
          type: string
        formed:
          type: string
        died:
          type: string
        disbanded:
          type: string
        biography:
          type: string
        path:
          type: string
          description: Filesystem path to the artist directory. Empty for pathless (platform-only) artists.
        library_id:
          type: string
        nfo_exists:
          type: boolean
        thumb_exists:
          type: boolean
        fanart_exists:
          type: boolean
        fanart_count:
          type: integer
        logo_exists:
          type: boolean
        banner_exists:
          type: boolean
        thumb_low_res:
          type: boolean
        fanart_low_res:
          type: boolean
        logo_low_res:
          type: boolean
        banner_low_res:
          type: boolean
        thumb_placeholder:
          type: string
          description: Base64-encoded placeholder for the thumb image. Omitted when not set.
        fanart_placeholder:
          type: string
          description: Base64-encoded placeholder for the fanart image. Omitted when not set.
        logo_placeholder:
          type: string
          description: Base64-encoded placeholder for the logo image. Omitted when not set.
        banner_placeholder:
          type: string
          description: Base64-encoded placeholder for the banner image. Omitted when not set.
        health_score:
          type: number
          format: double
        is_excluded:
          type: boolean
        exclusion_reason:
          type: string
          description: Reason the artist is excluded from rules and metadata operations. Omitted when not excluded.
        is_classical:
          type: boolean
        locked:
          type: boolean
          description: When true, automated operations will not modify this artist.
        lock_source:
          type: string
          description: Source that applied the lock (e.g. "user"). Omitted when not locked.
        locked_at:
          type: string
          format: date-time
          description: Timestamp when the lock was applied. Omitted when not locked.
        audiodb_id_fetched_at:
          type: string
          format: date-time
          description: When the AudioDB ID was last fetched. Omitted when never fetched.
        discogs_id_fetched_at:
          type: string
          format: date-time
          description: When the Discogs ID was last fetched. Omitted when never fetched.
        wikidata_id_fetched_at:
          type: string
          format: date-time
          description: When the Wikidata ID was last fetched. Omitted when never fetched.
        lastfm_fetched_at:
          type: string
          format: date-time
          description: When Last.fm metadata was last fetched. Omitted when never fetched.
        metadata_sources:
          type: object
          additionalProperties:
            type: string
          description: Map of field name to provider name indicating which provider supplied each field. Omitted when empty.
        last_scanned_at:
          type: string
          format: date-time
          description: When the artist directory was last scanned. Omitted when never scanned.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - sort_name
        - type
        - gender
        - disambiguation
        - musicbrainz_id
        - audiodb_id
        - discogs_id
        - wikidata_id
        - deezer_id
        - spotify_id
        - genres
        - styles
        - moods
        - years_active
        - born
        - formed
        - died
        - disbanded
        - biography
        - path
        - library_id
        - nfo_exists
        - thumb_exists
        - fanart_exists
        - fanart_count
        - logo_exists
        - banner_exists
        - thumb_low_res
        - fanart_low_res
        - logo_low_res
        - banner_low_res
        - health_score
        - is_excluded
        - is_classical
        - locked
        - created_at
        - updated_at
...
paths:
...
  /artists:
    get:
      tags: [Artists]
      summary: List artists
      operationId: listArtists
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: search
          in: query
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated artist list
          content:
            application/json:
              schema:
                type: object
                properties:
                  artists:
                    type: array
                    items:
                      $ref: "#/components/schemas/Artist"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
...
  /artists/{id}/aliases:
    get:
      tags: [Aliases]
      summary: List artist aliases
      operationId: listAliases
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of aliases
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    alias:
                      type: string
                    source:
                      type: string
    post:
      tags: [Aliases]
      summary: Add alias to artist
      operationId: addAlias
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alias:
                  type: string
                source:
                  type: string
              required: [alias]
      responses:
        "201":
          description: Alias created
          content:
            application/json:
              schema:
                type: object
...